requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
flask==2.3.3
pyyaml==6.0.1